Refactor: rename size to heuristicSize (ambiguity between Clause.size and Clause.lits.size was leading to bugs).
Change distance accumulation in waiting so that we add the log of the number of clauses
added to the set, not on a per clause basis by the number of its
literals (misread this in Hurd's). Otherwise, the empty clause gets a
distance of -inf and weight of NaN and so never gets pulled from the
sortedmap.

First set theory problem can now be resolved after 24 rounds.