Serious bug.
Hurd checks whether either of l or r are typedVars, and we dropped the
check since we don't care about types. However, an ordinary var is also
a typed var. Abbreviations require a variable on one side of the
equation!

Spotted in the following problem:

cnf(ax1,axiom,~vpair(A,B)=vpair(C,D) | B=D).

Here, vpair(A,B)=vpair(C,D) is taken as an abbreviation. Unification
says that B=D, and the clause is rewritten to D=D and then chucked out
as tautologous.